requirements.txt
feedparser==6.0.10
pyTelegramBotAPI==4.14.0
openai==1.40.0
gspread==6.1.2
google-auth==2.35.0
httpx==0.27.0